+++ spec/unit/packagers/mac_pkg_spec.rb
@@ -81,7 +81,8 @@ EOH
            package_dir: package_dir,
            package_tmp: package_tmp,
            mac_pkg_identifier: mac_pkg_identifier,
-           config: pkg_signing_config
+           config: pkg_signing_config,
+           friendly_name: 'Myproject'
   end
 
   let(:packager) do
@@ -152,7 +153,7 @@ EOH
     it 'generates the distribution and runs productbuild' do
       expect(packager).to receive(:execute).with ['productbuild',
                                                   %Q(--distribution "/var/cache/omnibus/pkg-tmp/mac_pkg/Distribution"),
-                                                  %Q(--resources "/omnibus/project/root/files/mac_pkg/Resources"),
+                                                  %Q(--resources "/var/cache/omnibus/pkg-tmp/mac_pkg/Resources"),
                                                   '/home/someuser/omnibus-myproject/pkg/myproject-23.4.2-4.pkg',
       ].join(' ')
       packager.build_product_pkg
@@ -170,7 +171,7 @@ EOH
     it 'includes the signing parameters in the product build command' do
       expect(packager).to receive(:execute).with ['productbuild',
                                                   %Q(--distribution "/var/cache/omnibus/pkg-tmp/mac_pkg/Distribution"),
-                                                  %Q(--resources "/omnibus/project/root/files/mac_pkg/Resources"),
+                                                  %Q(--resources "/var/cache/omnibus/pkg-tmp/mac_pkg/Resources"),
                                                   %Q(--sign "My Special Identity"),
                                                   '/home/someuser/omnibus-myproject/pkg/myproject-23.4.2-4.pkg',
       ].join(' ')